Example code to replace the an old PlatformKey with a new one.
Some platforms may have lost the private portion of the PlatformKey,
or the PlatformKey may have been compromised. In these cases, it is
necessary to replace the PlatformKey with a new one. This example
demonstrates how to replace the old PlatformKey with a new one. That
is intended to be used as a reference.

now targeting flat hash of cert instead of sig list of cert